Examples/Python/Containers/MongoDB-REST/requirements.txt
flask
prometheus_client
pymongo
python-dotenv
opentelemetry-api 
opentelemetry-sdk 
opentelemetry-instrumentation-flask 
opentelemetry-exporter-otlp
opentelemetry-exporter-otlp-proto-http
requests